Fix inconsistent usage of defines.
Either use
or

This cannot be mixed. If code defines SWITCH without a value the later
variant will fail, or default to zero, which is not the expected
behavior.
The first variant is used is most parts of the code, which makes more
sense anyway, because wit #ifdef there are only two possible optins:
defined and not defined.
The second approach has four possible state:
* defined
  - with 1 (on)
  - with 0 (off)
  - with other value (invalid)
* not defined (error)